+++ base/abstractdict.jl
@@ -719,10 +719,10 @@ Dict{Symbol,Int64} with 2 entries:
   :b => 2
 
 julia> map!(v -> v-1, values(d))
-Dict{Symbol,Int64} with 2 entries:
-  :a => 0
-  :b => 1
- ```
+Base.ValueIterator for a Dict{Symbol,Int64} with 2 entries. Values:
+  0
+  1
+```
 """
 function map!(f, iter::ValueIterator)
     # This is the naive fallback which requires hash evaluations