+++ stdlib/Profile/src/Profile.jl
@@ -5,7 +5,11 @@ Profiling support, main entry point is the [`@profile`](@ref) macro.
 """
 module Profile
 
-import Base.StackTraces: lookup, UNKNOWN, show_spec_linfo, StackFrame
+import Base.StackTraces: lookupat, UNKNOWN, show_spec_linfo, StackFrame
+
+# deprecated functions: use `getdict` instead
+lookup(ip::UInt) = lookupat(convert(Ptr{Cvoid}, ip) - 1)
+lookup(ip::Ptr{Cvoid}) = lookupat(ip - 1)
 
 export @profile
 
@@ -85,14 +89,16 @@ struct ProfileFormat
     sortedby::Symbol
     combine::Bool
     C::Bool
+    recur::Symbol
     function ProfileFormat(;
         C = false,
         combine = true,
         maxdepth::Int = typemax(Int),
         mincount::Int = 0,
         noisefloor = 0,
-        sortedby::Symbol = :filefuncline)
-        return new(maxdepth, mincount, noisefloor, sortedby, combine, C)
+        sortedby::Symbol = :filefuncline,
+        recur::Symbol = :off)
+        return new(maxdepth, mincount, noisefloor, sortedby, combine, C, recur)
     end
 end
 
@@ -115,37 +121,50 @@ The keyword arguments can be any combination of:
  - `maxdepth` -- Limits the depth higher than `maxdepth` in the `:tree` format.
 
  - `sortedby` -- Controls the order in `:flat` format. `:filefuncline` (default) sorts by the source
-    line, whereas `:count` sorts in order of number of collected samples.
+    line, `:count` sorts in order of number of collected samples, and `:overhead` sorts by the number of samples
+    incurred by each function by itself.
 
  - `noisefloor` -- Limits frames that exceed the heuristic noise floor of the sample (only applies to format `:tree`).
     A suggested value to try for this is 2.0 (the default is 0). This parameter hides samples for which `n <= noisefloor * √N`,
     where `n` is the number of samples on this line, and `N` is the number of samples for the callee.
 
  - `mincount` -- Limits the printout to only those lines with at least `mincount` occurrences.
+
+ - `recur` -- Controls the recursion handling in `:tree` format. `:off` (default) prints the tree as normal. `:flat` instead
+    compresses any recursion (by ip), showing the approximate effect of converting any self-recursion into an iterator.
+    `:flatc` does the same but also includes collapsing of C frames (may do odd things around `jl_apply`).
 """
-function print(io::IO, data::Vector{<:Unsigned} = fetch(), lidict::Union{LineInfoDict, LineInfoFlatDict} = getdict(data);
+function print(io::IO,
+        data::Vector{<:Unsigned} = fetch(),
+        lidict::Union{LineInfoDict, LineInfoFlatDict} = getdict(data)
+        ;
         format = :tree,
         C = false,
         combine = true,
         maxdepth::Int = typemax(Int),
         mincount::Int = 0,
         noisefloor = 0,
-        sortedby::Symbol = :filefuncline)
-    print(io, data, lidict, ProfileFormat(C = C,
+        sortedby::Symbol = :filefuncline,
+        recur::Symbol = :off)
+    print(io, data, lidict, ProfileFormat(
+            C = C,
             combine = combine,
             maxdepth = maxdepth,
             mincount = mincount,
             noisefloor = noisefloor,
-            sortedby = sortedby),
+            sortedby = sortedby,
+            recur = recur),
         format)
 end
 
 function print(io::IO, data::Vector{<:Unsigned}, lidict::Union{LineInfoDict, LineInfoFlatDict}, fmt::ProfileFormat, format::Symbol)
     cols::Int = Base.displaysize(io)[2]
     data = convert(Vector{UInt64}, data)
-    if format == :tree
+    fmt.recur ∈ (:off, :flat, :flatc) || throw(ArgumentError("recur value not recognized"))
+    if format === :tree
         tree(io, data, lidict, cols, fmt)
-    elseif format == :flat
+    elseif format === :flat
+        fmt.recur === :off || throw(ArgumentError("format flat only implements recur=:off"))
         flat(io, data, lidict, cols, fmt)
     else
         throw(ArgumentError("output format $(repr(format)) not recognized"))
@@ -174,12 +193,15 @@ allows you to save profiling results for future analysis.
 """
 function retrieve()
     data = fetch()
-    return (copy(data), getdict(data))
+    return (data, getdict(data))
 end
 
 function getdict(data::Vector{UInt})
-    uip = unique(data)
-    return LineInfoDict(UInt64(ip)=>lookup(ip) for ip in uip)
+    dict = LineInfoDict()
+    for ip in data
+        get!(() -> lookupat(convert(Ptr{Cvoid}, ip)), dict, UInt64(ip))
+    end
+    return dict
 end
 
 """
@@ -299,104 +321,113 @@ error_codes = Dict(
 """
     fetch() -> data
 
-Returns a reference to the internal buffer of backtraces. Note that subsequent operations,
-like [`clear`](@ref), can affect `data` unless you first make a copy. Note that the
+Returns a copy of the buffer of profile backtraces. Note that the
 values in `data` have meaning only on this machine in the current session, because it
 depends on the exact memory addresses used in JIT-compiling. This function is primarily for
 internal use; [`retrieve`](@ref) may be a better choice for most users.
 """
 function fetch()
-    len = len_data()
     maxlen = maxlen_data()
+    len = len_data()
     if (len == maxlen)
         @warn """The profile data buffer is full; profiling probably terminated
                  before your program finished. To profile for longer runs, call
                  `Profile.init()` with a larger buffer and/or larger delay."""
     end
-    return unsafe_wrap(Array, get_data_pointer(), (len,))
+    data = Vector{UInt}(undef, len)
+    GC.@preserve data unsafe_copyto!(pointer(data), get_data_pointer(), len)
+    # post-process the data to convert from a return-stack to a call-stack
+    first = true
+    for i = 1:length(data)
+        if data[i] == 0
+            first = true
+        elseif first
+            first = false
+        else
+            data[i] -= 1
+        end
+    end
+    return data
 end
 
 
 ## Print as a flat list
-# Counts the number of times each line appears, at any nesting level
-function count_flat(data::Vector{UInt64})
-    linecount = Dict{UInt64, Int}()
-    for ip in data
-        if ip != 0
-            linecount[ip] = get(linecount, ip, 0) + 1
-        end
-    end
-    iplist = Vector{UInt64}()
-    n = Vector{Int}()
-    for (k, v) in linecount
-        push!(iplist, k)
-        push!(n, v)
-    end
-    return (iplist, n)
-end
-
-function parse_flat(iplist::Vector{UInt64}, n::Vector{Int}, lidict::Union{LineInfoDict, LineInfoFlatDict}, C::Bool)
-    # Convert instruction pointers to names & line numbers
+# Counts the number of times each line appears, at any nesting level and at the topmost level
+# Merging multiple equivalent entries and recursive calls
+function parse_flat(::Type{T}, data::Vector{UInt64}, lidict::Union{LineInfoDict, LineInfoFlatDict}, C::Bool) where {T}
     lilist = StackFrame[]
-    nlist = Int[]
-    for (ip, count) in zip(iplist, n)
+    n = Int[]
+    m = Int[]
+    lilist_idx = Dict{T, Int}()
+    recursive = Set{T}()
+    first = true
+    totalshots = 0
+    for ip in data
+        if ip == 0
+            totalshots += 1
+            empty!(recursive)
+            first = true
+            continue
+        end
         frames = lidict[ip]
         nframes = (frames isa Vector ? length(frames) : 1)
-        # add all the inlining frames
-        for i = nframes:-1:1
+        for i = 1:nframes
             frame = (frames isa Vector ? frames[i] : frames)
-            # Keep only the interpretable ones
-            # The ones with no line number might appear multiple times in a single
-            # backtrace, giving the wrong impression about the total number of backtraces.
-            # Delete them too.
-            if frame != UNKNOWN && frame.line != 0 && (C || !frame.from_c)
+            !C && frame.from_c && continue
+            key = (T === UInt64 ? ip : frame)
+            idx = get!(lilist_idx, key, length(lilist) + 1)
+            if idx > length(lilist)
+                push!(recursive, key)
                 push!(lilist, frame)
-                push!(nlist, count)
+                push!(n, 1)
+                push!(m, 0)
+            elseif !(key in recursive)
+                push!(recursive, key)
+                n[idx] += 1
+            end
+            if first
+                m[idx] += 1
+                first = false
             end
         end
     end
-    return (lilist, nlist)
+    @assert length(lilist) == length(n) == length(m) == length(lilist_idx)
+    return (lilist, n, m, totalshots)
 end
 
 function flat(io::IO, data::Vector{UInt64}, lidict::Union{LineInfoDict, LineInfoFlatDict}, cols::Int, fmt::ProfileFormat)
-    iplist, n = count_flat(data)
-    lilist, n = parse_flat(iplist, n, lidict, fmt.C)
-    if isempty(n)
+    lilist, n, m, totalshots = parse_flat(fmt.combine ? StackFrame : UInt64, data, lidict, fmt.C)
+    if isempty(lilist)
         warning_empty()
         return
     end
-    print_flat(io, lilist, n, cols, fmt)
+    if false # optional: drop the "non-interpretable" ones
+        keep = map(frame -> frame != UNKNOWN && frame.line != 0, lilist)
+        lilist = lilist[keep]
+        n = n[keep]
+        m = m[keep]
+    end
+    print_flat(io, lilist, n, m, cols, fmt)
+    Base.println(io, "Total snapshots: ", totalshots)
     nothing
 end
 
-function print_flat(io::IO, lilist::Vector{StackFrame}, n::Vector{Int},
+function print_flat(io::IO, lilist::Vector{StackFrame},
+        n::Vector{Int}, m::Vector{Int},
         cols::Int, fmt::ProfileFormat)
-    p = liperm(lilist)
-    lilist = lilist[p]
-    n = n[p]
-    if fmt.combine
-        # now that lilist is sorted by li,
-        # combine adjacent entries that are equivlent
-        j = 1
-        for i = 2:length(lilist)
-            if lilist[i] == lilist[j]
-                n[j] += n[i]
-                n[i] = 0
-            else
-                j = i
-            end
-        end
-        keep = n .> 0
-        n = n[keep]
-        lilist = lilist[keep]
-    end
     if fmt.sortedby == :count
         p = sortperm(n)
-        n = n[p]
-        lilist = lilist[p]
+    elseif fmt.sortedby == :overhead
+        p = sortperm(m)
+    else
+        p = liperm(lilist)
     end
+    lilist = lilist[p]
+    n = n[p]
+    m = m[p]
     wcounts = max(6, ndigits(maximum(n)))
-    maxline = 0
+    wself = max(9, ndigits(maximum(m)))
+    maxline = 1
     maxfile = 6
     maxfunc = 10
     for li in lilist
@@ -405,28 +436,40 @@ function print_flat(io::IO, lilist::Vector{StackFrame}, n::Vector{Int},
         maxfunc = max(maxfunc, length(string(li.func)))
     end
     wline = max(5, ndigits(maxline))
-    ntext = cols - wcounts - wline - 3
-    maxfunc += 25
+    ntext = max(20, cols - wcounts - wself - wline - 3)
+    maxfunc += 25 # for type signatures
     if maxfile + maxfunc <= ntext
         wfile = maxfile
         wfunc = maxfunc
     else
-        wfile = floor(Integer, 2*ntext/5)
-        wfunc = floor(Integer, 3*ntext/5)
+        wfile = 2*ntext÷5
+        wfunc = 3*ntext÷5
     end
-    println(io, lpad("Count", wcounts, " "), " ", rpad("File", wfile, " "), " ",
-        lpad("Line", wline, " "), " ", rpad("Function", wfunc, " "))
+    println(io, lpad("Count", wcounts, " "), " ", lpad("Overhead", wself, " "), " ",
+            rpad("File", wfile, " "), " ", lpad("Line", wline, " "), " ", rpad("Function", wfunc, " "))
     for i = 1:length(n)
         n[i] < fmt.mincount && continue
         li = lilist[i]
         Base.print(io, lpad(string(n[i]), wcounts, " "), " ")
-        Base.print(io, rpad(rtruncto(string(li.file), wfile), wfile, " "), " ")
-        Base.print(io, lpad(string(li.line), wline, " "), " ")
-        fname = string(li.func)
-        if !li.from_c && li.linfo !== nothing
-            fname = sprint(show_spec_linfo, li)
+        Base.print(io, lpad(string(m[i]), wself, " "), " ")
+        if li == UNKNOWN
+            if !fmt.combine && li.pointer != 0
+                Base.print(io, "@0x", string(li.pointer, base=16))
+            else
+                Base.print(io, "[any unknown stackframes]")
+            end
+        else
+            file = string(li.file)
+            isempty(file) && (file = "[unknown file]")
+            Base.print(io, rpad(rtruncto(file, wfile), wfile, " "), " ")
+            Base.print(io, lpad(li.line > 0 ? string(li.line) : "?", wline, " "), " ")
+            fname = string(li.func)
+            if !li.from_c && li.linfo !== nothing
+                fname = sprint(show_spec_linfo, li)
+            end
+            isempty(fname) && (fname = "[unknown function]")
+            Base.print(io, rpad(ltruncto(fname, wfunc), wfunc, " "))
         end
-        Base.print(io, rpad(ltruncto(fname, wfunc), wfunc, " "))
         println(io)
     end
     nothing
@@ -435,13 +478,22 @@ end
 ## A tree representation
 tree_format_linewidth(x::StackFrame) = ndigits(x.line) + 6
 
-function tree_format(lilist::Vector{StackFrame}, counts::Vector{Int}, level::Int, cols::Int)
+const indent_s = "  ╎  "^10
+const indent_z = collect(eachindex(indent_s))
+function indent(depth::Int)
+    depth < 1 && return ""
+    depth <= length(indent_z) && return indent_s[1:indent_z[depth]]
+    div, rem = divrem(depth, length(indent_z))
+    return (indent_s^div) * SubString(indent_s, 1, indent_z[rem])
+end
+
+function tree_format(lilist::Vector{StackFrame}, counts::Vector{Int}, level::Int, cols::Int, showpointer::Bool)
     nindent = min(cols>>1, level)
     ndigcounts = ndigits(maximum(counts))
     ndigline = maximum([tree_format_linewidth(x) for x in lilist])
-    ntext = cols - nindent - ndigcounts - ndigline - 5
-    widthfile = floor(Integer, 0.4ntext)
-    widthfunc = floor(Integer, 0.6ntext)
+    ntext = max(20, cols - nindent - ndigcounts - ndigline - 5)
+    widthfile = 2*ntext÷5
+    widthfunc = 3*ntext÷5
     strs = Vector{String}(undef, length(lilist))
     showextra = false
     if level > nindent
@@ -452,7 +504,7 @@ function tree_format(lilist::Vector{StackFrame}, counts::Vector{Int}, level::Int
     for i = 1:length(lilist)
         li = lilist[i]
         if li != UNKNOWN
-            base = " "^nindent
+            base = nindent == 0 ? "" : indent(nindent - 1) * " "
             if showextra
                 base = string(base, "+", nextra, " ")
             end
@@ -464,9 +516,17 @@ function tree_format(lilist::Vector{StackFrame}, counts::Vector{Int}, level::Int
                     string(li.pointer, base = 16, pad = 2*sizeof(Ptr{Cvoid})),
                     ")")
             else
-                fname = string(li.func)
                 if !li.from_c && li.linfo !== nothing
                     fname = sprint(show_spec_linfo, li)
+                else
+                    fname = string(li.func)
+                end
+                if showpointer
+                    fname = string(
+                        "0x",
+                        string(li.pointer, base = 16, pad = 2*sizeof(Ptr{Cvoid})),
+                        " ",
+                        fname)
                 end
                 strs[i] = string(base,
                     rpad(string(counts[i]), ndigcounts, " "),
@@ -490,26 +550,63 @@ mutable struct StackFrameTree{T} # where T <: Union{UInt64, StackFrame}
     frame::StackFrame
     count::Int
     down::Dict{T, StackFrameTree{T}}
-    # construction helpers:
+    # construction workers:
+    recur::Bool
     builder_key::Vector{UInt64}
     builder_value::Vector{StackFrameTree{T}}
     up::StackFrameTree{T}
-    StackFrameTree{T}() where {T} = new(UNKNOWN, 0, Dict{T, StackFrameTree{T}}(), UInt64[], StackFrameTree{T}[])
+    StackFrameTree{T}() where {T} = new(UNKNOWN, 0, Dict{T, StackFrameTree{T}}(), false, UInt64[], StackFrameTree{T}[])
 end
 
 # turn a list of backtraces into a tree (implicitly separated by NULL markers)
-function tree!(root::StackFrameTree{T}, all::Vector{UInt64}, lidict::Union{LineInfoFlatDict, LineInfoDict}, C::Bool) where {T}
+function tree!(root::StackFrameTree{T}, all::Vector{UInt64}, lidict::Union{LineInfoFlatDict, LineInfoDict}, C::Bool, recur::Symbol) where {T}
     parent = root
-    for i in length(all):-1:1
+    tops = Vector{StackFrameTree{T}}()
+    build = Vector{StackFrameTree{T}}()
+    startframe = length(all)
+    for i in startframe:-1:1
         ip = all[i]
         if ip == 0
             # sentinel value indicates the start of a new backtrace
+            empty!(build)
+            if recur !== :off
+                # We mark all visited nodes to so we'll only count those branches
+                # once for each backtrace. Reset that now for the next backtrace.
+                push!(tops, parent)
+                for top in tops
+                    while top.recur
+                        top.recur = false
+                        top = top.up
+                    end
+                end
+                empty!(tops)
+            end
             parent = root
             parent.count += 1
+            startframe = i
         else
+            pushfirst!(build, parent)
+            if recur === :flat || recur == :flatc
+                # Rewind the `parent` tree back, if this exact ip was already present *higher* in the current tree
+                found = false
+                for j in 1:(startframe - i)
+                    if ip == all[i + j]
+                        if recur === :flat # if not flattening C frames, check that now
+                            frames = lidict[ip]
+                            frame = (frames isa Vector ? frames[1] : frames)
+                            frame.from_c && break
+                        end
+                        push!(tops, parent)
+                        parent = build[j]
+                        found = true
+                        break
+                    end
+                end
+                found && continue
+            end
             builder_key = parent.builder_key
             builder_value = parent.builder_value
-            fastkey = searchsortedfirst(parent.builder_key, ip)
+            fastkey = searchsortedfirst(builder_key, ip)
             if fastkey < length(builder_key) && builder_key[fastkey] === ip
                 # jump forward to the end of the inlining chain
                 # avoiding an extra (slow) lookup of `ip` in `lidict`
@@ -517,9 +614,12 @@ function tree!(root::StackFrameTree{T}, all::Vector{UInt64}, lidict::Union{LineI
                 # note that we may even have this === parent (if we're ignoring this frame ip)
                 this = builder_value[fastkey]
                 let this = this
-                    while this !== parent
-                        this.count += 1
-                        this = this.up
+                    if recur === :off || !this.recur
+                        while this !== parent && !this.recur
+                            this.count += 1
+                            this.recur = true
+                            this = this.up
+                        end
                     end
                 end
                 parent = this
@@ -533,12 +633,13 @@ function tree!(root::StackFrameTree{T}, all::Vector{UInt64}, lidict::Union{LineI
                 frame = (frames isa Vector ? frames[i] : frames)
                 !C && frame.from_c && continue
                 key = (T === UInt64 ? ip : frame)
-                this = get!(parent.down, key) do
-                    return StackFrameTree{T}()
+                this = get!(StackFrameTree{T}, parent.down, key)
+                if recur === :off || !this.recur
+                    this.frame = frame
+                    this.up = parent
+                    this.count += 1
+                    this.recur = true
                 end
-                this.frame = frame
-                this.up = parent
-                this.count += 1
                 parent = this
             end
             # record where the end of this chain is for this ip
@@ -547,9 +648,10 @@ function tree!(root::StackFrameTree{T}, all::Vector{UInt64}, lidict::Union{LineI
         end
     end
     function cleanup!(node::StackFrameTree)
-        stack = StackFrameTree[node]
+        stack = [node]
         while !isempty(stack)
             node = pop!(stack)
+            node.recur = false
             empty!(node.builder_key)
             empty!(node.builder_value)
             append!(stack, values(node.down))
@@ -562,7 +664,7 @@ end
 
 # Print the stack frame tree starting at a particular root. Uses a worklist to
 # avoid stack overflows.
-function tree(io::IO, bt::StackFrameTree, cols::Int, fmt::ProfileFormat)
+function print_tree(io::IO, bt::StackFrameTree{T}, cols::Int, fmt::ProfileFormat) where T
     worklist = [(bt, 0, 0, "")]
     while !isempty(worklist)
         (bt, level, noisefloor, str) = popfirst!(worklist)
@@ -574,7 +676,7 @@ function tree(io::IO, bt::StackFrameTree, cols::Int, fmt::ProfileFormat)
         lilist = collect(frame.frame for frame in nexts)
         counts = collect(frame.count for frame in nexts)
         # Generate the string for each line
-        strs = tree_format(lilist, counts, level, cols)
+        strs = tree_format(lilist, counts, level, cols, T === UInt64)
         # Recurse to the next level
         for i in reverse(liperm(lilist))
             down = nexts[i]
@@ -591,15 +693,16 @@ end
 
 function tree(io::IO, data::Vector{UInt64}, lidict::Union{LineInfoFlatDict, LineInfoDict}, cols::Int, fmt::ProfileFormat)
     if fmt.combine
-        root = tree!(StackFrameTree{StackFrame}(), data, lidict, fmt.C)
+        root = tree!(StackFrameTree{StackFrame}(), data, lidict, fmt.C, fmt.recur)
     else
-        root = tree!(StackFrameTree{UInt64}(), data, lidict, fmt.C)
+        root = tree!(StackFrameTree{UInt64}(), data, lidict, fmt.C, fmt.recur)
     end
     if isempty(root.down)
         warning_empty()
         return
     end
-    tree(io, root, cols, fmt)
+    print_tree(io, root, cols, fmt)
+    Base.println(io, "Total snapshots: ", root.count)
     nothing
 end